feat: Implement platform settings management and enhance billing features
- Introduced PlatformSettingsService to manage global operator settings.
- Added migration for platform_settings table to store settings persistently.
- Updated PlatformBillingService to utilize settings for invoice due dates and prefixes.
- Enhanced API routes to include endpoints for managing platform settings and analytics.
- Created tests for platform settings to ensure correct behavior and validation.
- Implemented idempotency key in mikrotik_sync_logs for provisioning operations.
- Added retry backoff configuration for provisioning jobs.

+++ app/Http/Controllers/Api/PlatformSubscriptionController.php
@@ -59,11 +59,11 @@ class PlatformSubscriptionController extends Controller
             'cancelled_subscriptions' => TenantSubscription::where('status', 'cancelled')->count(),
             // MRR = PrimeBill's recurring revenue from tenants: active monthly
             // prices + active annual prices amortized at 1/12. ARR = MRR × 12.
-            // Kept consistent with PlatformAdminService::getOverviewStats.
-            'mrr' => TenantSubscription::where('status', 'active')
+            // Consistent with PlatformAdminService::getOverviewStats.
+            'mrr' => (float) TenantSubscription::where('status', 'active')
                 ->where('billing_cycle', 'monthly')
                 ->sum('price')
-                + TenantSubscription::where('status', 'active')
+                + (float) TenantSubscription::where('status', 'active')
                     ->where('billing_cycle', 'annual')
                     ->sum('price') / 12,
         ];
